+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ############################ canic ####################################
 
+## [0.2.3] - 2025-10-15
+- app_directory and subnet_directory are now calculated from the SubnetCanisterRegistry
+- directories are now part of CanisterInitPayload, with the Env struct, sent to a canister as its created
+
 ## [0.2.2] - 2025-10-13
 - removed all the delegation code
 - added in ops::signature, a wrapper around creating and verifying canister signatures